feat: 구조변경 및 임시커밋
shelf 실패로 인한 임시커밋임.
이 상태로 되돌리면 안됨.
