+++ src/main/java/com/aluminium/acoe/app/admin/persistance/AdminCafeRepositoryCustom.java
@@ -1,7 +1,7 @@
 package com.aluminium.acoe.app.admin.persistance;
 
 import com.aluminium.acoe.app.admin.dto.AdminCafeSearchDto;
-import com.aluminium.acoe.app.cafe.entity.Cafe;
+import com.aluminium.acoe.app.main.entity.Cafe;
 import java.util.List;
 
 public interface AdminCafeRepositoryCustom {
