+++ src/main/java/com/aluminium/acoe/app/batch/service/BatchService.java
@@ -1,0 +1,8 @@
+package com.aluminium.acoe.app.batch.service;
+
+public interface BatchService {
+    /**
+     * 카페 정보 조회 일배치
+     */
+    void executeCafeBatch();
+}
