--- src/main/java/com/aluminium/acoe/app/cafe/entity/Menu.java
@@ -1,32 +1,0 @@
-package com.aluminium.acoe.app.cafe.entity;
-
-import com.aluminium.acoe.common.entity.BaseEntity;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Table(name = "menu")
-@Getter
-@Builder
-@AllArgsConstructor(access = AccessLevel.PROTECTED)
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Menu extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long menuId;
-
-    @Column(name = "menu_nm", length = 50)
-    private String menuNm;
-
-    @Column(name = "price")
-    private Long price;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "cafe_id", updatable = false)
-    private Cafe cafe;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "franchise_id", updatable = false)
-    private Franchise franchise;
-}
