--- src/main/java/com/aluminium/acoe/app/cafe/service/BatchService.java
@@ -1,8 +1,0 @@
-package com.aluminium.acoe.app.cafe.service;
-
-public interface BatchService {
-    /**
-     * 카페 정보 조회 일배치
-     */
-    void executeCafeBatch();
-}
